requirements.txt
# FicTGNN: Financial Technology Graph Neural Networks
# Requirements for academic research in financial network analysis

# Core dependencies
torch>=1.12.0
torch-geometric>=2.1.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Data processing
requests>=2.25.0
lxml>=4.6.0
beautifulsoup4>=4.9.0

# Visualization
plotly>=5.0.0
dash>=2.0.0
networkx>=2.6.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
pyyaml>=6.0
psutil>=5.8.0
tqdm>=4.62.0
jupyter>=1.0.0

# Optional GPU support
# torch-scatter>=2.0.9
# torch-sparse>=0.6.12
# torch-cluster>=1.5.9
# torch-spline-conv>=1.2.1

# Development dependencies
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0